requirements.txt
tqdm
python-dotenv
openai~=1.67.0
python-decouple~=3.8